# Backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Web App Template API",
    "description": "Backend API for the web app template.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "List users",
        "description": "Return all users.",
        "operationId": "list_users",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserRead"
                  },
                  "type": "array",
                  "title": "Response List Users"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create a user",
        "description": "Register a new user. The password is hashed before storage.",
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get a user",
        "description": "Return a single user by id.",
        "operationId": "get_user",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exercises": {
      "get": {
        "tags": [
          "exercises"
        ],
        "summary": "List exercises",
        "description": "Return a page of the current user's exercises.",
        "operationId": "list_exercises",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (1-indexed).",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (1-indexed)."
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of items per page.",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Number of items per page."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ExerciseRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "exercises"
        ],
        "summary": "Create an exercise",
        "description": "Create a new exercise owned by the current user.",
        "operationId": "create_exercise",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exercises/{exercise_id}": {
      "patch": {
        "tags": [
          "exercises"
        ],
        "summary": "Update an exercise",
        "description": "Update an exercise owned by the current user. Omitted fields are left unchanged.",
        "operationId": "update_exercise",
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Exercise Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "exercises"
        ],
        "summary": "Delete an exercise",
        "description": "Delete an exercise owned by the current user.",
        "operationId": "delete_exercise",
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Exercise Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workout-templates": {
      "get": {
        "tags": [
          "workout-templates"
        ],
        "summary": "List workout templates",
        "description": "Return all of the current user's workout templates.",
        "operationId": "list_workout_templates",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkoutTemplateRead"
                  },
                  "type": "array",
                  "title": "Response List Workout Templates"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workout-templates"
        ],
        "summary": "Create a workout template",
        "description": "Create a new workout template owned by the current user.",
        "operationId": "create_workout_template",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutTemplateCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutTemplateRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workout-templates/{template_id}": {
      "get": {
        "tags": [
          "workout-templates"
        ],
        "summary": "Get a workout template",
        "description": "Return a single workout template owned by the current user.",
        "operationId": "get_workout_template",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutTemplateRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "workout-templates"
        ],
        "summary": "Update a workout template",
        "description": "Update a workout template owned by the current user. Omitted fields are left unchanged.",
        "operationId": "update_workout_template",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Template Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutTemplateUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutTemplateRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "workout-templates"
        ],
        "summary": "Delete a workout template",
        "description": "Delete a workout template owned by the current user.",
        "operationId": "delete_workout_template",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ExerciseCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Exercise name.",
            "examples": [
              "Bench Press"
            ]
          },
          "muscle_group": {
            "$ref": "#/components/schemas/MuscleGroup",
            "description": "Primary muscle group targeted."
          }
        },
        "type": "object",
        "required": [
          "name",
          "muscle_group"
        ],
        "title": "ExerciseCreate",
        "description": "Payload for creating a new exercise."
      },
      "ExerciseRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique identifier."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Exercise name.",
            "examples": [
              "Bench Press"
            ]
          },
          "muscle_group": {
            "$ref": "#/components/schemas/MuscleGroup",
            "description": "Primary muscle group targeted."
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By",
            "description": "Id of the user who created this exercise, if any."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "When the exercise was created."
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "muscle_group",
          "created_at"
        ],
        "title": "ExerciseRead",
        "description": "An exercise as returned by the API."
      },
      "ExerciseUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Exercise name.",
            "examples": [
              "Bench Press"
            ]
          },
          "muscle_group": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MuscleGroup"
              },
              {
                "type": "null"
              }
            ],
            "description": "Primary muscle group targeted."
          }
        },
        "type": "object",
        "title": "ExerciseUpdate",
        "description": "Payload for updating an exercise. Omitted fields are left unchanged."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MuscleGroup": {
        "type": "string",
        "enum": [
          "chest",
          "back",
          "shoulders",
          "biceps",
          "triceps",
          "abs",
          "calves",
          "quads"
        ],
        "title": "MuscleGroup"
      },
      "Page_ExerciseRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExerciseRead"
            },
            "type": "array",
            "title": "Items",
            "description": "The items on this page."
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of items across all pages."
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "Current page number (1-indexed).",
            "examples": [
              1
            ]
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size",
            "description": "Maximum number of items per page.",
            "examples": [
              20
            ]
          },
          "pages": {
            "type": "integer",
            "title": "Pages",
            "description": "Total number of pages."
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages"
        ],
        "title": "Page[ExerciseRead]"
      },
...
      "TemplateExerciseCreate": {
        "properties": {
          "exercise_id": {
            "type": "string",
            "format": "uuid",
            "title": "Exercise Id",
            "description": "Id of the exercise to add."
          },
          "set_count": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Set Count",
            "description": "Number of planned sets.",
            "examples": [
              3
            ]
          }
        },
        "type": "object",
        "required": [
          "exercise_id",
          "set_count"
        ],
        "title": "TemplateExerciseCreate",
        "description": "Payload for one exercise slot within a new workout template."
      },
      "TemplateExerciseRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique identifier."
          },
          "exercise_id": {
            "type": "string",
            "format": "uuid",
            "title": "Exercise Id",
            "description": "Id of the exercise in this slot."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Exercise name.",
            "examples": [
              "Bench Press"
            ]
          },
          "muscle_group": {
            "$ref": "#/components/schemas/MuscleGroup",
            "description": "Primary muscle group targeted."
          },
          "position": {
            "type": "integer",
            "title": "Position",
            "description": "Display order within the template, 1-indexed."
          },
          "set_count": {
            "type": "integer",
            "title": "Set Count",
            "description": "Number of planned sets."
          }
        },
        "type": "object",
        "required": [
          "id",
          "exercise_id",
          "name",
          "muscle_group",
          "position",
          "set_count"
        ],
        "title": "TemplateExerciseRead",
        "description": "An exercise slot within a workout template, as returned by the API."
      },
...
      "UserCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "User email address.",
            "examples": [
              "user@example.com"
            ]
          },
          "first_name": {
            "type": "string",
            "maxLength": 100,
            "title": "First Name",
            "description": "Given name.",
            "examples": [
              "Jane"
            ]
          },
          "last_name": {
            "type": "string",
            "maxLength": 100,
            "title": "Last Name",
            "description": "Family name.",
            "examples": [
              "Doe"
            ]
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "description": "Plaintext password (min 8 characters)."
          }
        },
        "type": "object",
        "required": [
          "email",
          "first_name",
          "last_name",
          "password"
        ],
        "title": "UserCreate",
        "description": "Payload for creating a user. The password is stored hashed, never in plaintext."
      },
...
      "UserRead": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "User email address.",
            "examples": [
              "user@example.com"
            ]
          },
          "first_name": {
            "type": "string",
            "maxLength": 100,
            "title": "First Name",
            "description": "Given name.",
            "examples": [
              "Jane"
            ]
          },
          "last_name": {
            "type": "string",
            "maxLength": 100,
            "title": "Last Name",
            "description": "Family name.",
            "examples": [
              "Doe"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique identifier."
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Whether the account is active."
          },
          "is_superuser": {
            "type": "boolean",
            "title": "Is Superuser",
            "description": "Whether the account has elevated privileges."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "When the account was created."
          }
        },
        "type": "object",
        "required": [
          "email",
          "first_name",
          "last_name",
          "id",
          "is_active",
          "is_superuser",
          "created_at"
        ],
        "title": "UserRead",
        "description": "A user as returned by the API. Never includes the password hash."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkoutTemplateCreate": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "description": "Template name.",
            "examples": [
              "Push Day"
            ]
          },
          "exercises": {
            "items": {
              "$ref": "#/components/schemas/TemplateExerciseCreate"
            },
            "type": "array",
            "minItems": 1,
            "title": "Exercises",
            "description": "Exercise slots to add, in order."
          }
        },
        "type": "object",
        "required": [
          "name",
          "exercises"
        ],
        "title": "WorkoutTemplateCreate",
        "description": "Payload for creating a new workout template."
      },
      "WorkoutTemplateRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique identifier."
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id",
            "description": "Id of the user who owns this template."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Template name.",
            "examples": [
              "Push Day"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "When the template was created."
          },
          "exercises": {
            "items": {
              "$ref": "#/components/schemas/TemplateExerciseRead"
            },
            "type": "array",
            "title": "Exercises",
            "description": "Exercise slots, in order."
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "name",
          "created_at",
          "exercises"
        ],
        "title": "WorkoutTemplateRead",
        "description": "A workout template as returned by the API."
      },
      "WorkoutTemplateUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Template name.",
            "examples": [
              "Push Day"
            ]
          },
          "exercises": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TemplateExerciseCreate"
                },
                "type": "array",
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Exercises",
            "description": "Replacement exercise slots, in order."
          }
        },
        "type": "object",
        "title": "WorkoutTemplateUpdate",
        "description": "Payload for updating a workout template. Omitted fields are left unchanged.\n\nWhen ``exercises`` is provided, it replaces the entire set of slots;\nposition is assigned from list order, so reordering is done by\nresubmitting the list in the desired order."
      }
    }
  }
}